Changes to reset.css, index.php, and login.php
Added citation for reset.css.

Included the functionality to print out user input from login.php (to
make sure it was sending) and also to get closer to sending the data to
the database.

login.php uses index.php as the action now and also will send the data
to post on submit.